Remove the MESSAGE.Solaris and INSTALL.openwin.
The INSTALL.openwin is broken.

I reported on December 1 that it create a broken INSTALL script.

The SunOS check was commented out in pkgsrc/x11/xterm/Makefile revision
1.37. It looks like it was accidently committed as it has nothing to do
with the log message.

This caused the INSTALL.openwin stuff to be used on non-SunOS systems.

That INSTALL.openwin support is broken anyways.  The sed changed @PREFIX@
but that was not used.