Update nkf to 2.08b aka 2.0.8b.
pkgsrc changes:
add DESTDIR support

Changes 2.0.6
== Addition of --ic --oc option
--Ic=<input codeset> --oc=<output codeset> option was added.
Shift_JIS, CP932, EUC-JP, CP51932, eucJP-ms, ISO-2022-JP, UTF-8N(UTF-8),
UTF-8-BOM, UTF-16BE, UTF-16BE-BOM(UTF-16), UTF-16LE, and UTF-16LE-BOM.
UTF8-MAC can be specified for input.

== Addition and change in mapping
The character-code of JIS-affiliated and correspondence with Unicode
decided to be defined. Decode from these to Unicode secures the thing
that conversion in the defined range is correctly done. Conversion in
the range not defined is not secured. Encode from Unicode secures
the thing that conversion in the defined range is correctly done,
conversion in the range not defined is assumed to be an error,
(possible fallback), and Unicode, and an illegal character secures
the skipped thing.

== Addition of fallback option
The --fb-{skip,html,xml,perl,java,subchar} option was added. How the
character that was not able to be converted is treated when
converting it from Unicode into Shift_JIS, EUC-JP, and ISO-2022-JP
can be specified.

== Addition of --no-best-fit-chars option
The character for which the round trip safety is not secured is not
converted at conversion from Unicode. - when Unicode is converted
from Unicode It is possible to use it to convert the UTF form of nkf
when specifying it with x. (If this is not specified if it says
oppositely, a part of character is not preserved. )

When the character string related to passing is converted, this
option is strongly recommended to be specified.

== Addition of --in-place[=SUFFIX] option
--in-place option to renew the time stamp to --overwrite that
maintained the time stamp was added.

== Document of hiragana katakana conversion
Because the document of the hiragana katakana conversion had reversed,
this was corrected. A correct document is as follows.
    -H1    --hiragana katakana > hiragana
    -H2    --katakana hiragana -> katakana
    -H3    --katakana-hiragana katakana <-> hiragana

== Others
"Q" decipherment of encoding and the bug such as fold-margin are corrected.

== 2.0.6a
The permission of the file at the packaging was changed.

Changes 2.0.7
* The character of CP51932 corrects not doing fallback.
* It is corrected not to compile excluding gcc.
* Making to ANSI C.
* The bug around changing line of MIME encode is corrected.
* The binary file came to be also good at base64 by -t2MB.
* Support the conversion of repeating hiragana and katakana sign by -h.
* Support the conversion of V of the hiragana and the katakana by -h.
* DOS and OS/2 1.x support by lukewarm
* OpenWatcom, OS/2, and djgpp V2 support by By lukewarm

Changes 2.0.8
* Support of UTF-32 and CP10001(Microsoft's MacJapanese).
* It supports outside BMP of Unicode.
* The user definition character of the CP932 faction is supported.
* The range of the Shift_JIS emoticon of Softbank Mobile is supported with NTT DoCoMo.
* The line feed code was displayed with --guess.
* It is corrected that the guess when the character-code cannot be fixed with --guess was broken.
* It is corrected to have thrown away SI/SO/ESC at the EUC mode.
* When the input of UTF-8 is encoded in MIME, it is corrected not to be output
correctly.
* Processing was continued even if there was a file that was not able to be read
when two or more files were given.
* Various corrections in CP932 turn

In 2.0.8a in addition to above
* It is corrected not effective of -L.
* It is corrected not to be able to make nkf32.dll.