Import gate88-20050319 as games/gate88.
Gate 88 is often described as "like subspace but with the ability
to build turrets and command fighters." While a skillful pilot can
decimate the enemy in one-on-one combat, a thoughtful player can
overwhelm the opponent with strategic turret and fighter placements.
However, only those who truly understand the intricate details of
both action and strategy will win the day.

Because alliances can be broken on a whim, players must interpret
the social situation of the game to ensure that each alliance is
sincere. For example, it is not uncommon for an allied player to
plant turrets and fighters within your base only to break the
alliance and watch your empire crumble from within. Always be
careful with your trust.

Gate 88 employs a unique user interface that allows players to
pilot their ship while performing strategic operations. The tutorial
and practice mode will get you up to speed.