Behold:  The JACK audio connection kit version 0.109.10 (SVN). This update was long overdue from version 0.100.0.  Especially of note: native audio(4) support (OSS driver initially converted by jakemsr@openbsd.org), with certain changes/fixups by me (both in the driver code and elsewhere ... it is very much still a work in progress, but it is a work that does work, heh).
Also, some patches to do the "right thing" throughout, i.e. press the
"we are not linux" button.  More fixes/enhancements coming soon.

One outstanding issue I know of:  an mmap failure when trying to run
with realtime scheduling that I haven't been able to look into enough yet.

A good player which supports this is audio/moc ... I hope to add
a gstreamer plugin and much more shortly.

I am in the planning stages of a rewrite of the driver using kqueue, also.